Dev pow

+++ types/spend.go
@@ -1,3 +1,7 @@
+// Copyright 2020 The Reed Developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
 package types
 
 import (
@@ -8,7 +12,7 @@ import (
 type Spend struct {
 	SpendOutputId Hash   `json:"spendOutputId"`
 	SoureId       Hash   `json:"-"` //rel utxo.SourceId
-	SourcePos     uint32 `json:"-"` //rel utxo.SourcePos
+	SourcePos     uint64 `json:"-"` //rel utxo.SourcePos
 	Amount        uint64 `json:"-"` //rel utxo.Amount
 	ScriptPk      []byte `json:"-"` //rel utxo.ScriptPK
 }
